Add jobTitle and company as user inputs to CLI
When users signup via Modulus CLI they will be prompted for a job
title and a company. These fields are optional. Tested this by
registering a user and signing in via xammr:55443. This version of
aperture is connected to Congress via port 8686. User is added to
the database with the associated fields.

[MODULUS-1150]