Add `mod-project-name` to package.json
When deploying with --project-name flag fulcrum will now check
to see if the `mod-project-name` object is in the package.json.
If it is not then it will add it and rewrite the file in a neat
order. It will append either at the top of the file or after the
`name` object.

[MODULUS-1332]